feat(uishell): make component based
