+++ src/components/UIShell/SideNav/SideNavItems.svelte
@@ -1,0 +1,7 @@
+<script>
+  import { cx } from '../../../lib';
+</script>
+
+<ul class={cx('--side-nav__items')}>
+  <slot />
+</ul>
