+++ src/components/UIShell/index.js
@@ -1,3 +1,9 @@
 import UIShell from './UIShell.svelte';
 
 export default UIShell;
+
+export { default as SideNav } from './SideNav/SideNav.svelte';
+export { default as SideNavItems } from './SideNav/SideNavItems.svelte';
+export { default as SideNavLink } from './SideNav/SideNavLink.svelte';
+export { default as SideNavMenu } from './SideNav/SideNavMenu.svelte';
+export { default as SideNavMenuItem } from './SideNav/SideNavMenuItem.svelte';
